+++ 2026spring/frontend/src/app/page.tsx
@@ -3,8 +3,12 @@ import Hero from "@/components/Hero";
 import About from "@/components/About";
 import Rules from "@/components/Rules";
 import Schedule from "@/components/Schedule";
-import LinksAndSponsors from "@/components/LinksAndSponsors";
-import PastEventLink from "@/components/PastEventLink";
+
+import RelatedLinks from "@/components/sections/RelatedLinks";
+import Crowdfunding from "@/components/sections/Crowdfunding";
+import PastEventLink from "@/components/sections/PastEventLink";
+import Sponsors from "@/components/sections/Sponsors";
+import Contact from "@/components/sections/Contact";
 
 
 export default function Home() {
@@ -14,8 +18,12 @@ export default function Home() {
       <About />
       <Rules />
       <Schedule />
-      <LinksAndSponsors />
+
+      <RelatedLinks />
+      <Crowdfunding />
       <PastEventLink />
+      <Sponsors />
+      <Contact />
 
       <footer className="w-full py-8 text-center text-sm text-slate-400 bg-white">
         <p>© 2026 本当のルーキー祭り 運営</p>
